Core: renames component to Core
