+++ src/components/Core/index.js
@@ -1,4 +1,4 @@
-import Core from './src/core';
+import Core from './src/Core';
 
 Core.install = function(Vue) {
   Vue.component(Core.name, Core);
